+++ javascope/jScope/JetDataProvider.java
@@ -146,7 +146,7 @@ class JetDataProvider implements DataProvider
          }
 
         public float[] getZ(){System.out.println("BADABUM!!"); return null;}
-        public float[] getX2D(){System.out.println("BADABUM!!"); return null;}
+        public double[] getX2D(){System.out.println("BADABUM!!"); return null;}
         public long[] getX2DLong(){System.out.println("BADABUM!!"); return null;}
         public float[] getY2D(){System.out.println("BADABUM!!"); return null;} 
         public double[] getXLimits(){System.out.println("BADABUM!!"); return null;}
